feat: Add LevelBuilder, and related classes,also refactor BlockData namespace
