+++ Assets/Scripts/Level/ILevelDesigner.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c1fdfc5a653fc6b46ac2072f68581899
