+++ Assets/Scripts/Level/LevelDesigner.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 364f3e40e23117d48bc62823579a3451
